requirements.txt
certifi==2018.1.18
chardet==3.0.4
coverage==4.4.2
coveralls==1.2.0
Django==2.0.1
django-cors-headers==2.2.0
django-nose==1.4.5
djangorestframework==3.7.7
docopt==0.6.2
gunicorn==19.7.1
idna==2.6
nose==1.3.7
numpy==1.14.2
pytz==2018.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
sklearn==0.0
urllib3==1.22
whitenoise==3.3.1
